git-svn-id: svn://svn.icms.temple.edu/lammps-ro/trunk@14164 f3b2605a-c512-4ea7-a41b-209d697bcdaa

+++ src/USER-MISC/fix_ti_rs.h
@@ -53,9 +53,9 @@ class FixTIRS : public Fix {
   double l_initial;    // Lambda initial value.
   double l_final;      // Lambda final value.
   double linfo[2];     // Current lambda status.
-  int    t_switch;     // Total switching steps.
-  int    t_equil;      // Equilibration time.
-  int    t0;           // Initial time.
+  bigint t_switch;     // Total switching steps.
+  bigint t_equil;      // Equilibration time.
+  bigint t0;           // Initial time.
   int    sf;           // Switching function option.
   int    nlevels_respa;
 };
